refactor: split GTK toolbar views into focused modules
